Last Commit

+++ app/Http/Controllers/profile_controller.php
@@ -20,7 +20,7 @@ class profile_controller extends Controller
     }
 
     public function update(Request $request){
-       // print_r($request->all());
+
 
         $request->validate(
             [
@@ -31,7 +31,6 @@ class profile_controller extends Controller
                 'name.required' => 'Name is required',
                 'phone_number.required' => 'Phone number is required',
                 'phone_number.size' => 'Phone number must be 11 digits',
-                //'phone_number.numeric' => 'Phone number must be numeric',
             ]
             );
            User::where('id',Auth::user()->id)->update([
@@ -39,23 +38,26 @@ class profile_controller extends Controller
             'phone_number' => $request->phone_number,
             'address' => $request->address,
             'updated_at' => Carbon::now(),
-           ]);// id check kore
+           ]);
+           if($request->hasFile('profile_image')){
+            if(Auth::user()->profile_image != '0.jpg'){
+                unlink(public_path('uploads/profile_image/'.Auth::user()->profile_image));
+                }
 
-           if(Auth::user()->profile_image != '0.jpg'){
-            unlink(public_path('uploads/profile_image/'.Auth::user()->profile_image));
-            }
+               $id=Auth::user()->id;
+               $extention=$request->profile_image->getClientOriginalExtension();
+               $file_name=$id.".".$extention;
 
-           $id=Auth::user()->id;
-           $extention=$request->profile_image->getClientOriginalExtension();
-           $file_name=$id.".".$extention;
+               $request->profile_image->move(public_path('uploads/profile_image'),$file_name);
+               User::where('id',Auth::user()->id)->update([
+                'profile_image'=>$file_name,
+            ]);
+           }
 
-           $request->profile_image->move(public_path('uploads/profile_image'),$file_name);
-           User::where('id',Auth::user()->id)->update([
-            'profile_image'=>$file_name,
-        ]);
+
 
            return back()->with('success','Profile Updated Successfully');
-          // return redirect()->route('profile_update')->with('success','Profile Updated Successfully');
+          return redirect()->route('profile_update')->with('success','Profile Updated Successfully');
 
 
     }
